[gd] update the benchmarks, and add a reference to verify the device results

+++ src/gd-cuda/utils.cpp
@@ -57,7 +57,7 @@ void get_CRSM_from_svm(Classification_Data_CRS &M, const std::string &file_path)
   }
   else {
     std::cout << "Could not find the SMV file, check again!" << std::endl;
-    exit(1); 
+    exit(1);
   }
   //numRows will be given by the rowpointer size -1
   M.m = static_cast<int>(M.row_ptr.size())-1;
@@ -73,7 +73,7 @@ void get_CRSM_from_svm(Classification_Data_CRS &M, const std::string &file_path)
     //Step 1):
     auto norm_sqrd = 0.0;
     for(auto j = M.row_ptr[i]; j < M.row_ptr[i+1]; j++){
-      assert(j < M.values.size());
+      assert((size_t)j < M.values.size());
       norm_sqrd += std::pow(M.values[j], 2);
     }
     auto norm = std::sqrt(norm_sqrd);
